requirements.txt
torch>=2.2
torchvision>=0.17
accelerate>=0.28
transformers>=4.41
diffusers>=0.29
opencv-python>=4.8
matplotlib>=3.7
numpy>=1.24
tqdm>=4.66
Pillow>=10.0
geopandas>=0.14
shapely>=2.0
pyproj>=3.6
fiona>=1.9
